fix: handle vet token recovery in coin class
This is because wrw calls coin class' recover function instead of token class
recover function

Ticket: COIN-6042
